// ai/macro/develop/review/types.ts
export type ReviewMetric = string |
  "security" |
  "performance" |
  "maintainability" |
  "reliability" |
  "readability" |
  "testability" |
  "portability" |
  "usability" |
  "accessibility" |
  "concurrency" |
  "internationalization" |
  "localization" |
  "documentation" |
  "efficiency" |
  "correctness" |
  "robustness" |
  "flexibility" |
  "extensibility" |
  "reusability" |
  "interoperability";

/**
 * Detected framework score.
 */
export interface DetectedFrameworks {
  [key: Framework]: number;
}

/**
 * Metric weighted values
 */
export interface ReviewMetricValues {
  weight: number;
  target: number;
  description: string;
}

/**
 * The specifications for each metric
 */
export interface MetricSpecifications {
  [key: ReviewMetric]: ReviewMetricValues;
}